requirements.txt
fastapi==0.110.1
uvicorn==0.29.0
requests==2.31.0
python-dotenv==1.0.1
transformers==4.40.1
torch==2.3.0
sqlalchemy==2.0.30
pydantic==1.10.14
